roles_drf/requirement.txt
annotated-types==0.5.0
asgiref==3.7.2
bleach==6.1.0
crispy-bootstrap4==2023.1
Django==4.2.4
django-bootstrap-datepicker==1.5.0
django-bootstrap-datepicker-plus==5.0.4
django-crispy-forms==2.1
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.0.0
pydantic==2.2.1
pydantic_core==2.6.1
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
webencodings==0.5.1